# apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Devbox API",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "def-0": {
        "type": "object",
        "required": [
          "id",
          "name",
          "image",
          "status",
          "containerId",
          "networkName",
          "volumeName",
          "tailnetUrl",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "creating"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "running"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "stopping"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "stopped"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "orphaned"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "removing"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "error"
                ]
              }
            ]
          },
          "containerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "networkName": {
            "type": "string"
          },
          "volumeName": {
            "type": "string"
          },
          "tailnetUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "deletedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "Box"
      },
      "def-1": {
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "boxId",
          "progress",
          "message",
          "error",
          "createdAt",
          "startedAt",
          "finishedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "create"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "stop"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "remove"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "sync"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "cleanup"
                ]
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "queued"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "running"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "succeeded"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "failed"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "cancelled"
                ]
              }
            ]
          },
          "boxId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "Job"
      }
    }
  },
  "paths": {
    "/openapi.json": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/boxes": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "command": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "env": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "box",
                    "job"
                  ],
                  "properties": {
                    "box": {
                      "$ref": "#/components/schemas/def-0"
                    },
                    "job": {
                      "$ref": "#/components/schemas/def-1"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/def-0"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/boxes/{boxId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "boxId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "boxId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/v1/boxes/{boxId}/stop": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "boxId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/v1/boxes/{boxId}/logs": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "follow",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "since",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "boxId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/jobs": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/def-1"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/jobs/{jobId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "jobId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-1"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}